fix: pytestconfig
