+++ tests/v0/conftest.py
@@ -1,6 +1,5 @@
 # Copyright 2023 Canonical Ltd.
 # See LICENSE file for licensing details.
-import argparse
 import os
 from importlib.metadata import version
 from unittest.mock import PropertyMock
@@ -8,32 +7,6 @@ from unittest.mock import PropertyMock
 import pytest
 from ops import JujuVersion
 from pytest_mock import MockerFixture
-
-
-def pytest_addoption(parser):
-    parser.addoption(
-        "--os-series", help="Ubuntu series for dp libs charm (e.g. jammy)", default="jammy"
-    )
-    parser.addoption(
-        "--build-bases-index",
-        type=int,
-        help="Index of charmcraft.yaml base that matches --os-series",
-        default=0,
-    )
-
-
-def pytest_configure(config):
-    if (config.option.os_series is None) ^ (config.option.build_bases_index is None):
-        raise argparse.ArgumentError(
-            None,
-            "--os-series and --build-bases-index must be given together",
-        )
-    # Note: Update defaults whenever charmcraft.yaml is changed
-    valid_combinations = [(0, "jammy"), (1, "focal")]
-    if (config.option.build_bases_index, config.option.os_series) not in valid_combinations:
-        raise argparse.ArgumentError(
-            None, f"Only base index combinations {valid_combinations} are accepted."
-        )
 
 
 @pytest.fixture(autouse=True)
